Don't just keep going when opening a git URL fails
Once we identified that our filename is actually a git designator (as seen
by the fact that it ends in a [branchname] surrounded by '[]'), we
shouldn't try to open that filename in order to try other ways of parsing
the data; instead we should just return an error to the caller.

This way the calling code can tell that an error occured.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>